music4you/requirements.txt
coverage==4.5.1
coveralls==1.3.0
Django==2.0.7
django-nose==1.4.5
djangorestframework==3.8.2
Pillow==5.2.0